addfile

+++ C0120G1/Module2/src/Demo/Ep_Kieu/Main_2.java
@@ -6,7 +6,7 @@ public class Main_2 {
     public static void main(String[] args) {
         int a=5;
         float b;
-        b=(int)a;
+        b=a;
         System.out.println(b);
 
         float c=6.5f;
